Fix integer type in test

+++ src/torrent/picker/sequential.rs
@@ -107,7 +107,7 @@ mod tests {
         };
 
         let picker = Picker::new(&info);
-        assert_eq!(picker.c.scale as usize, info.piece_len / 16384);
+        assert_eq!(picker.c.scale as u32, info.piece_len / 16384);
         assert_eq!(picker.c.blocks.len(), 123);
     }
